tod_system/requirement.txt
absl-py==2.0.0
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.2
bidict==0.22.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
datasets==2.15.0
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
dnspython==2.4.2
evaluate==0.4.1
eventlet==0.34.2
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
fsspec==2023.10.0
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.4
llama_cpp_python==0.2.23
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.4.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rouge-score==0.1.2
sacrebleu==2.4.0
safetensors==0.4.1
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tabulate==0.9.0
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
